disha:change in product module

+++ app/Http/Controllers/Backend/ProductController.php
@@ -193,12 +193,15 @@ class ProductController extends MainController
         $product = Product::where('id',$id)->delete();
         if($product){
             $image_info = ProductImage::where('product_id',$id)->first();
-            $image = $image_info->image;
-            $product_id = $image_info->product_id;
-            if($image){
-                removeFile('uploads/product/'.$product_id.'/'.$image);
+            if(isset($image_info))
+            {
+                $image = $image_info->image;
+                $product_id = $image_info->product_id;
+                if($image){
+                    removeFile('uploads/product/'.$product_id.'/'.$image);
+                }
+                $product_image = ProductImage::where('product_id',$id)->delete();
             }
-            $product_image = ProductImage::where('product_id',$id)->delete();
             return redirect()->back()->with('success', trans('Product Deleted Successfully!'));
         } else {
             return redirect()->back()->with('error', trans('Something went wrong, please try again later!'));
